# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REST API Pedro",
        "description": "REST API -> TCC of Pedro Rodrigues Pereira",
        "contact": {
            "email": "jierlausson@hotmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sub-activities": {
            "get": {
                "tags": [
                    "SubActivities"
                ],
                "summary": "List all sub-activities",
                "operationId": "9c61a0e984fb2ad44fa5a1afa3a2bf9c",
                "responses": {
                    "200": {
                        "description": "List of sub-activities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SubActivity"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "SubActivities"
                ],
                "summary": "Create a new sub-activity",
                "operationId": "b0453c1dabe0909b264eb89b8bb016b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "activity_id": {
                                        "type": "integer"
                                    },
                                    "class_room_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "workload_max": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "workload": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sub-activity created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubActivity"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sub-activities/{subActivity}": {
            "get": {
                "tags": [
                    "SubActivities"
                ],
                "summary": "Retrieve a specific sub-activity",
                "operationId": "710d2995ffe7c9a03001cd3cc6dfb3a8",
                "parameters": [
                    {
                        "name": "subActivity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sub-activity details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubActivity"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "SubActivities"
                ],
                "summary": "Update a sub-activity",
                "operationId": "9cfd7132011ae67f3bfa25e9098e3d9b",
                "parameters": [
                    {
                        "name": "subActivity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "activity_id": {
                                        "type": "integer"
                                    },
                                    "class_room_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "workload_max": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "workload": {
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sub-activity updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubActivity"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "SubActivities"
                ],
                "summary": "Delete a sub-activity",
                "operationId": "367e0cd61d7eed349f8fa6237df544ab",
                "parameters": [
                    {
                        "name": "subActivity",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sub-activity deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SubActivity": {
                "title": "SubActivity",
                "description": "Modelo da subatividade",
                "required": [
                    "id",
                    "activity_id",
                    "description",
                    "workload_max"
                ],
                "properties": {
                    "id": {
                        "description": "ID da subatividade",
                        "type": "integer"
                    },
                    "activity_id": {
                        "description": "ID da atividade associada",
                        "type": "integer"
                    },
                    "class_room_id": {
                        "description": "ID da sala de aula associada",
                        "type": "integer",
                        "nullable": true
                    },
                    "description": {
                        "description": "Descrição da subatividade",
                        "type": "string"
                    },
                    "workload_max": {
                        "description": "Carga horária máxima",
                        "type": "integer"
                    },
                    "workload": {
                        "description": "Carga horária atual",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Data de atualização",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}